requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdf2image==1.17.0
pillow==10.1.0
pydantic==2.5.0
httpx==0.25.2